% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bubbleSort}
\alias{bubbleSort}
\title{Title bubbleSort}
\usage{
bubbleSort(vector)
}
\value{
vector
}
\description{
This is a function to sort out an out of order sequence
}
\examples{
vector = c(5,34,65,36,67,3,6,43,69,59,25,785,10,11,14)
         bubbleSort(vector)
          [1]   3   5   6  10  11  14  25  34  36  43  59  65  67  69 785
}
