% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggman.R
\name{ggman}
\alias{ggman}
\title{Manhattan plot based on ggplot2}
\usage{
ggman(
  x,
  chr = "CHR",
  bp = "POS",
  p = "P",
  col = c("#1f78b4", "#a6cee3"),
  chrlabs = c(1:10, 12, 14, 17, 20, 24, 36),
  sigline = NULL,
  logp = TRUE,
  xlab = "Chromosome",
  ylab = "P",
  title = NULL,
  plot = TRUE,
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{A data.frame.}

\item{chr}{Chromosome for the horizontal axis}

\item{bp}{Base pair position}

\item{p}{Statistics for the vertical axis}

\item{col}{Vector of color codes of points}

\item{chrlabs}{Labels of chromosomes to be printed}

\item{sigline}{Numeric value for a significant line}

\item{logp}{Boolean to use log-scaled y-axis}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{title}{Title of the plot}

\item{plot}{Boolean to show plot}

\item{ylim}{Ranges of y-axis}

\item{...}{Other arguments passed on to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}}
}
\description{
Manhattan plot based on ggplot2
}
