% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{theme_my}
\alias{theme_my}
\alias{common_theme}
\alias{facet_theme}
\title{Custom theme for ggplot2}
\usage{
theme_my(
  theme = "bw",
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

common_theme()

facet_theme()
}
\arguments{
\item{theme}{"bw" or "test". Specifies theme_*().}

\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
\code{theme_my()} slightly modifies \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw()}} or \code{\link[ggplot2:ggtheme]{ggplot2::theme_test()}}

Common settings for \code{\link[ggplot2:theme]{ggplot2::theme()}} and myrrr ggtheme

Custom settings for \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} and \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}
}
