% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtreemix.R
\name{read_treemix}
\alias{read_treemix}
\alias{plot_treemix}
\alias{theme_treemix}
\title{A light-weighted handler of TreeMix result for ggplot2
These functions are slightly modified version of \code{treemix}
from \code{andrewparkermorgan/popcorn}}
\usage{
read_treemix(stem, ...)

plot_treemix(
  obj,
  plot.nodes = FALSE,
  plot.migration = TRUE,
  branch.colour = "#333333",
  branch.width = 0.5,
  label = TRUE,
  ...
)

theme_treemix(...)
}
\arguments{
\item{stem}{filename stem (\code{-o} from command-line) for TreeMix output}

\item{...}{passed to \link[ggplot2:ggtheme]{ggplot2::theme_bw}.}

\item{obj}{treemix result parsed by \code{read_treemix()}}

\item{plot.nodes}{Whether to plot nodes}

\item{plot.migration}{Whether to plot migration edges}

\item{branch.colour}{Color of tree branch}

\item{branch.width}{Linewidth of tree branch}

\item{label}{Whether to show tip names}
}
\description{
A light-weighted handler of TreeMix result for ggplot2
These functions are slightly modified version of \code{treemix}
from \code{andrewparkermorgan/popcorn}

Plot treemix result using \code{ggplot2}

Original treemix like plot theme
}
