% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grcg7b}
\alias{grcg7b}
\title{GRCg7b.}
\format{
A data frame with two variables: \code{CHROM}, \code{chr}
}
\source{
\url{https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/016/699/485/GCF_016699485.2_bGalGal1.mat.broiler.GRCg7b/GCF_016699485.2_bGalGal1.mat.broiler.GRCg7b_assembly_report.txt}
}
\usage{
grcg7b
}
\description{
Sequence report for GRCg7b.
}
\keyword{datasets}
