% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrrr-package.R
\docType{package}
\name{myrrr-package}
\alias{myrrr}
\alias{myrrr-package}
\title{myrrr: Personal R package}
\description{
Personal R package.
}
\author{
\strong{Maintainer}: Yuki Matsuda \email{yuki.matsuda.m10@gmail.com} (\href{https://orcid.org/0009-0006-2288-1621}{ORCID})

}
\keyword{internal}
