% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyplots_colors.R
\docType{data}
\name{colors_discrete_r3}
\alias{colors_discrete_r3}
\title{Custom color schemes of tidyplots}
\format{
An object of class \code{tidycolor} (inherits from \code{character}) of length 8.
}
\usage{
colors_discrete_r3
}
\description{
Custom color schemes of tidyplots
}
\examples{
(energy \%>\%
   tidyplot(year, power, color = energy_source) \%>\%
   add_barstack_absolute() \%>\%
   adjust_colors(new_colors = colors_discrete_set1)
}
\keyword{datasets}
