% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Annotate gene name to bed style dataframe}
\usage{
annotate(df, seqnames, start, end, colname = NULL)
}
\arguments{
\item{df}{A data.frame. with bed style}

\item{seqnames}{Colname passed to seqnames.field}

\item{start}{Colname passed to start.field}

\item{end}{Colname passed to end.field}

\item{colname}{EctraColname to be printed}
}
\description{
Annotate gene name to bed style dataframe
}
