% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyplots_colors.R
\docType{data}
\name{colors_discrete_r3}
\alias{colors_discrete_r3}
\alias{colors_discrete_r4}
\alias{colors_discrete_accent}
\alias{colors_discrete_dark2}
\alias{colors_discrete_paired}
\alias{colors_discrete_set1}
\alias{colors_discrete_set2}
\alias{colors_discrete_set3}
\title{Custom color schemes of tidyplots}
\format{
An object of class \code{tidycolor} (inherits from \code{character}) of length 8.

An object of class \code{tidycolor} (inherits from \code{character}) of length 8.

An object of class \code{tidycolor} (inherits from \code{character}) of length 8.

An object of class \code{tidycolor} (inherits from \code{character}) of length 8.

An object of class \code{tidycolor} (inherits from \code{character}) of length 12.

An object of class \code{tidycolor} (inherits from \code{character}) of length 9.

An object of class \code{tidycolor} (inherits from \code{character}) of length 8.

An object of class \code{tidycolor} (inherits from \code{character}) of length 12.
}
\usage{
colors_discrete_r3

colors_discrete_r4

colors_discrete_accent

colors_discrete_dark2

colors_discrete_paired

colors_discrete_set1

colors_discrete_set2

colors_discrete_set3
}
\description{
Custom discrete color schemes. \verb{See palette.pals()}.
}
\keyword{datasets}
